small refactoring, updated Dockefile
