+++ src/tg_bot_float_db_updater/main.py
@@ -1,7 +1,7 @@
 from typing import List
 from fastapi import FastAPI
 
-from tg_bot_float_abstracts.router_controller.abstract_router_controller import (
+from tg_bot_float_misc.router_controller.abstract_router_controller import (
     AbstractRouterController,
 )
 from tg_bot_float_db_updater.router_controllers.db_data_updater_router import (
